+++ lib/src/typedefs.dart
@@ -3,7 +3,7 @@ import 'package:flutter/widgets.dart';
 /// A callback function that takes a [BuildContext] and a state of type [T]
 /// and returns a [Future] that completes when the action on the state is
 /// completed.
-typedef OnSuccess<T, S> = Future<void> Function(BuildContext context, T state);
+typedef OnSuccess<T, S> = void Function(BuildContext context, T state);
 
 /// A callback function that takes a state of type [S] and performs an action
 /// on that state.